pyproject.toml
[tool.poetry]
name = "energy-api"
version = "0.1.0"
description = "This will interact with the kraken api so as to be abb information."
authors = ["Awen Saunders <me@awen.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gql = {extras = ["aiohttp", "httpx"], version = "^3.5.0"}
httpx = "^0.26.0"
pydantic = "^2.6.1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
ariadne-codegen = {extras = ["subscriptions"], version = "^0.12.0"}
debugpy = "^1.8.1"

[tool.ariadne-codegen]
queries_path = "gql/"
# remote_schema_url = "https://api.eonnext-kraken.energy/v1/graphql/"
schema_path = "schema.gql"
plugins = ["ariadne_codegen.contrib.shorter_results.ShorterResultsPlugin"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
